Report fixes
